[WIP] fix compile error
