+++ examples/local_rdma_ops/write_gpu.cpp
@@ -13,8 +13,8 @@ using namespace mori::core;
       IBV_ACCESS_REMOTE_ATOMIC
 
 template <ProviderType P>
-__device__ void SendThreadKernel(RdmaEndpoint& epSend, MemoryRegion sendMr, RdmaEndpoint& epRecv,
-                                 MemoryRegion recvMr, int msgSize, int msgNum) {
+__device__ void SendThreadKernel(RdmaEndpoint& epSend, RdmaMemoryRegion sendMr, RdmaEndpoint& epRecv,
+                                 RdmaMemoryRegion recvMr, int msgSize, int msgNum) {
     uint8_t sendVal = msgNum;
     for (int j = 0; j < msgSize; j++) {
       reinterpret_cast<char*>(sendMr.addr)[j] = sendVal;
@@ -43,7 +43,7 @@ __device__ void SendThreadKernel(RdmaEndpoint& epSend, MemoryRegion sendMr, Rdma
 }
 
 template <ProviderType P>
-__device__ void RecvThreadKernel(RdmaEndpoint& epRecv, MemoryRegion mr, int msgSize, int msgNum) {
+__device__ void RecvThreadKernel(RdmaEndpoint& epRecv, RdmaMemoryRegion mr, int msgSize, int msgNum) {
     printf("round %d recv verify for result\n", msgNum);
     uint8_t sendVal = msgNum;
 
@@ -58,8 +58,8 @@ __device__ void RecvThreadKernel(RdmaEndpoint& epRecv, MemoryRegion mr, int msgS
            reinterpret_cast<uint8_t*>(mr.addr)[768]);
 }
 
-__global__ void SendRecvOnGpu(RdmaEndpoint& epSend, RdmaEndpoint& epRecv, MemoryRegion mrSend,
-                              MemoryRegion mrRecv, int msgSize, int msgNum) {
+__global__ void SendRecvOnGpu(RdmaEndpoint& epSend, RdmaEndpoint& epRecv, RdmaMemoryRegion mrSend,
+                              RdmaMemoryRegion mrRecv, int msgSize, int msgNum) {
   assert(gridDim.x == 1);
   int bid = blockIdx.x;
   for (int i = 0; i < msgNum; i++) {
@@ -98,7 +98,7 @@ void LocalRdmaOps() {
 
   // RDMA initialization
   // 1 Create device
-  RdmaContext rdmaContext;
+  RdmaContext rdmaContext(RdmaBackendType::DirectVerbs);
   printf("nums_device: %d\n", rdmaContext.nums_device);
   RdmaDeviceList rdmaDevices = rdmaContext.GetRdmaDeviceList();
   assert(!rdmaDevices.empty());
@@ -138,18 +138,18 @@ void LocalRdmaOps() {
   HIP_RUNTIME_CHECK(hipMemcpy(devEpRecv, &epRecv, sizeof(RdmaEndpoint), hipMemcpyHostToDevice));
   void* sendBuf;
   HIP_RUNTIME_CHECK(hipMalloc(&sendBuf, msgSize));
-  MemoryRegion mrSend = deviceContextSend->RegisterMemoryRegion(sendBuf, msgSize, MR_ACCESS_FLAG);
+  RdmaMemoryRegion mrSend = deviceContextSend->RegisterRdmaMemoryRegion(sendBuf, msgSize, MR_ACCESS_FLAG);
 
   void* recvBuf;
   HIP_RUNTIME_CHECK(hipMalloc(&recvBuf, msgSize));
-  MemoryRegion mrRecv = deviceContextRecv->RegisterMemoryRegion(recvBuf, msgSize, MR_ACCESS_FLAG);
+  RdmaMemoryRegion mrRecv = deviceContextRecv->RegisterRdmaMemoryRegion(recvBuf, msgSize, MR_ACCESS_FLAG);
 
   SendRecvOnGpu<<<1, 1>>>(*devEpSend, *devEpRecv, mrSend, mrRecv, msgSize, msgNum);
   HIP_RUNTIME_CHECK(hipDeviceSynchronize());
 
   // 8 Finalize
-  deviceContextSend->DeRegisterMemoryRegion(sendBuf);
-  deviceContextRecv->DeRegisterMemoryRegion(recvBuf);
+  deviceContextSend->DeregisterRdmaMemoryRegion(sendBuf);
+  deviceContextRecv->DeregisterRdmaMemoryRegion(recvBuf);
   HIP_RUNTIME_CHECK(hipFree(devEpSend));
   HIP_RUNTIME_CHECK(hipFree(devEpRecv));
   HIP_RUNTIME_CHECK(hipFree(sendBuf));
