+++ include/mori/application/transport/rdma/rdma.hpp
@@ -31,7 +31,8 @@
 #include <vector>
 
 #include "infiniband/verbs.h"
-#include "mori/core/transport/rdma/primitives.hpp"
+#include "mori/core/transport/rdma/rdma.hpp"
+// #include "mori/core/transport/rdma/primitives.hpp"
 
 namespace mori {
 namespace application {
@@ -64,7 +65,7 @@ RdmaDeviceVendorId ToRdmaDeviceVendorId(T v) {
 }
 
 #define PAGESIZE uint32_t(sysconf(_SC_PAGE_SIZE))
-#define OUTSTANDING_TABLE_SIZE (65536)
+// #define OUTSTANDING_TABLE_SIZE (65536)
 
 /* -------------------------------------------------------------------------- */
 /*                             Rdma Data Structure                            */
@@ -124,65 +125,20 @@ struct WorkQueueAttrs {
   uint32_t offset{0};
 };
 
-struct WorkQueueHandle {
-  uint32_t postIdx{0};     // numbers of wqe that post to work queue
-  uint32_t dbTouchIdx{0};  // numbers of wqe that touched doorbell
-  uint32_t doneIdx{0};     // numbers of wqe that have been consumed by nic
-  uint32_t readyIdx{0};
-  union {
-    struct {
-      uint32_t msntblSlotIdx;
-      uint32_t psnIdx; // for bnxt msn psn index calculate
-    };
-    uint64_t msnPack{0};
-  };
-  void* sqAddr{nullptr};
-  void* msntblAddr{nullptr}; // for bnxt
-  void* rqAddr{nullptr};
-  void* dbrRecAddr{nullptr};
-  void* dbrAddr{nullptr};
-  uint32_t mtuSize{4096};
-  uint32_t sqWqeNum{0};
-  uint32_t msntblNum{0};
-  uint32_t rqWqeNum{0};
-  uint32_t postSendLock{0};
-  uint64_t outstandingWqe[OUTSTANDING_TABLE_SIZE]{0};
-};
-
-struct CompletionQueueHandle {
-  void* cqAddr{nullptr};
-  void* dbrRecAddr{nullptr};
-  void* dbrAddr{nullptr};
-  uint32_t consIdx{0}; // numbers of cqe that have been completed
-  uint32_t needConsIdx{0}; // numbers of cqe that should be consumed
-  uint32_t activeIdx{0}; // numbers of cqe that under processing but not completed
-  uint32_t cq_consumer{0};
-  uint32_t cqeNum{0};
-  uint32_t cqeSize{0};
-  uint32_t pollCqLock{0};
-};
-
-struct IBVerbsHandle {
-  ibv_qp* qp{nullptr};
-  ibv_cq* cq{nullptr};
-  ibv_srq* srq{nullptr};
-  ibv_comp_channel* compCh{nullptr};
-};
-
 struct RdmaEndpoint {
   RdmaDeviceVendorId vendorId{RdmaDeviceVendorId::Unknown};
   RdmaEndpointHandle handle;
   // TODO(@ditian12): we should use an opaque handle to reference the actual transport context
   // handles, for direct verbs it should be WorkQueueHandle/CompletionQueueHandle, for ib verbs, it
   // should be ibv structures
-  WorkQueueHandle wqHandle;
-  CompletionQueueHandle cqHandle;
-  IBVerbsHandle ibvHandle;
+  core::WorkQueueHandle wqHandle;
+  core::CompletionQueueHandle cqHandle;
+  core::IBVerbsHandle ibvHandle;
 
   __device__ __host__ core::ProviderType GetProviderType() {
     if (vendorId == RdmaDeviceVendorId::Mellanox) {
       return core::ProviderType::MLX5;
-    } if (vendorId == Broadcom) {
+    } if (vendorId == RdmaDeviceVendorId::Broadcom) {
       return core::ProviderType::BNXT;
     } else {
       printf("unknown vendorId %d", vendorId);
