+++ src/io/rdma/backend_impl_v1.cpp
@@ -200,6 +200,7 @@ std::vector<std::pair<int, int>> RdmaManager::Search(TopoKey key) {
   } else {
     assert("topo searching for device other than GPU is not implemented yet");
   }
+  return {};
 }
 
 /* ----------------------------------- Local Memory Management ---------------------------------- */
@@ -377,7 +378,7 @@ void NotifManager::RegisterDevice(int devId) {
 }
 
 void NotifManager::MainLoop() {
-  int maxEvents = 128;
+  constexpr int maxEvents = 128;
   epoll_event events[maxEvents];
   while (running.load()) {
     int nfds = epoll_wait(epfd, events, maxEvents, 0 /*ms*/);
@@ -619,7 +620,7 @@ void ControlPlaneServer::HandleControlPlaneProtocol(int fd) {
 }
 
 void ControlPlaneServer::MainLoop() {
-  int maxEvents = 128;
+  constexpr int maxEvents = 128;
   epoll_event events[maxEvents];
 
   while (running.load()) {
